Update ProjectE to 1.4.1

+++ src/main/java/com/latmod/mods/projectex/ProjectEXCommon.java
@@ -15,7 +15,7 @@ public class ProjectEXCommon
 		return null;
 	}
 
-	public void updateEMC(double emc)
+	public void updateEMC(long emc)
 	{
 	}
 }
